added Ultraservo helper and basic ultrasonic codes

+++ FtcRobotController/src/main/java/lib/DriveTrain.java
@@ -7,6 +7,6 @@ public interface DriveTrain {
     // DriveTrain describes the drive setup of the robot.
     // This can be four wheel drive, two wheel drive, treads, etc, so we need to implement
     // this interface in other classes.
-    public void move(float power);
+    public void move(double power);
     public void turnWithEncoders(float power, int degrees);
 }
